rover_listener: resolved NameError exception
The reason things were blowing up was that there was a variable 'conn' that was
assigned to in each loop. The expression on the right hand side waited for data
to be sent to the server. If there was an interrupt during that time then
essentially 'conn' was never assigned to, and the 'finally' block had an
undeclared variable causing the 'NameError' exception to be raised.